Added IEEE SaTML 2025
IEEE SaTML is a conference focused on the intersection of machine learning and security/privacy. It expands on the theoretical and practical understandings of vulnerabilities inherent to machine learning (ML), explores the robustness of learning algorithms, and aids in developing a unified, coherent scientific community which aims to establish trustworthy machine learning.